add topology pipeline tasks demo
