+++ packages/react-topology/src/components/TopologyView/examples/topology-example.css
@@ -3,6 +3,7 @@
 .ws-react-t-custom-nodes,
 .ws-react-t-getting-started,
 .ws-react-t-panzoom,
+.ws-react-t-pipelines,
 .ws-react-t-selection,
 .ws-react-t-sidebar,
 .ws-react-t-toolbar {
